llmops-article-classifier/requirements.txt
accelerate>=0.26.0
anthropic==0.46.0
beautifulsoup4==4.13.3
colorama==0.4.6
datasets==3.2.0
evaluate==0.4.3
huggingface_hub==0.28.1
numpy==2.2.3
pandas==2.2.3
polars==1.22.0
psutil==6.1.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
scikit_learn==1.6.1
tenacity==9.0.0
torch==2.6.0
tqdm>=4.64.0
transformers==4.48.0.dev0
typing_extensions==4.12.2
zenml==0.75.0